Load server config from file and tidy menu
