+++ src/scripts/start.js
@@ -1,9 +1,9 @@
 /* globals $ */
 
 import { remote, ipcRenderer } from 'electron';
-import { servers } from './servers';
-import { sidebar } from './sidebar';
-import { webview } from './webview';
+import servers from './servers';
+import sidebar from './sidebar';
+import webview from './webview';
 import tray from './tray';
 import './menus';
